requirements/base.txt
Django==1.6.1
Jinja2==2.7.1
MarkupSafe==0.18
Pygments==1.6
South==0.8.4
redis==2.9.1
anyjson==0.3.3
billiard==3.3.0.13
celery==3.1.6
django-browserid==0.9
django-cors-headers==0.12
django-model-utils==1.5.0
djangorestframework==2.3.10
docutils==0.11
fancy-tag==0.2.0
kombu==3.0.8
parse==1.6.3
prody==1.5.1
psycopg2==2.5.1
pytz==2013.8
requests==2.1.0
six==1.5.2
sqlparse==0.1.10
wsgiref==0.1.2
